feat: Finished setup for google and apple authentication.
What I did: Finished authentication main, and google and apple. Have to test and rebuild.

Why it matters: With authentication mostly done, we can now move on to the onboarding screens of the real app once authentication is tested.

Next: Test and rebuild to test deep links and authentication with google and apple. Move on to onboarding screens if good, otherwise, reiterate.